# api/openapi.yaml
openapi: 3.0.3
info:
  title: authsvc API
  version: 0.3.0
  description: Plug-and-play authentication and authorization service
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    BootstrapSecret:
      type: http
      scheme: bearer
      description: Bootstrap secret as Bearer token for initial admin setup
  schemas:
    TokenResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer }
        refresh_token: { type: string, nullable: true }
        scope: { type: string }
        id_token: { type: string, nullable: true }
    AuthzCheckRequest:
      type: object
      required: [subject_id, tenant_id, resource, action]
      properties:
        subject_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        resource: { type: string }
        action: { type: string }
        context: { type: object, additionalProperties: true }
    AuthzResult:
      type: object
      properties:
        allowed: { type: boolean }
        reason: { type: string, nullable: true }
    ApiKeyCreateRequest:
      type: object
      required: [tenant_id, name, scopes]
      properties:
        tenant_id: { type: string, format: uuid }
        name: { type: string }
        scopes: { type: array, items: { type: string } }
paths:
...
  /oauth/token:
    post:
      summary: OAuth2 token endpoint
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [grant_type]
              properties:
                grant_type:
                  type: string
                  enum:
                    - password
                    - refresh_token
                    - client_credentials
                    - authorization_code
                    - api_key
                client_id: { type: string }
                client_secret: { type: string }
                username: { type: string }
                password: { type: string }
                refresh_token: { type: string }
                code: { type: string }
                redirect_uri: { type: string }
                code_verifier: { type: string }
                api_key: { type: string }
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
  /oauth/userinfo:
    get:
      summary: OIDC userinfo
      security:
        - BearerAuth: []
      responses:
        "200": { description: User claims }
...
  /oauth/authorize:
    get:
      summary: Authorization code flow (PKCE)
      parameters:
        - { name: client_id, in: query, required: true, schema: { type: string } }
        - { name: redirect_uri, in: query, required: true, schema: { type: string } }
        - { name: response_type, in: query, required: true, schema: { type: string } }
        - { name: code_challenge, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: Login state for browser flow }
...
  /oauth/federate/{provider}:
    get:
      summary: Start federated login (Google, GitHub)
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Authorization URL }
...
  /v1/auth/register:
    post:
      summary: Register user (bootstrap or admin)
      security:
        - BootstrapSecret: []
      responses:
        "200": { description: User created }
  /v1/clients:
    post:
      summary: Create OAuth client (admin)
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "200": { description: Client created }
    get:
      summary: List OAuth clients (admin)
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "200": { description: Client list }
  /v1/authz/check:
    post:
      summary: Authorization check (RBAC, Casbin, or OpenFGA)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthzCheckRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthzResult"
  /v1/mfa/enroll:
    post:
      summary: Enroll TOTP MFA
      security:
        - BearerAuth: []
      responses:
        "200": { description: OTPAuth URL and recovery codes }
...
  /v1/api-keys:
    post:
      summary: Create API key
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyCreateRequest"
      responses:
        "200": { description: API key created }
  /v1/keys/rotate:
    post:
      summary: Rotate JWT signing keys
      security:
        - BearerAuth: []
      responses:
        "200": { description: New kid }
  /v1/roles/{child_id}/inherit/{parent_id}:
    post:
      summary: Link child role to parent role for permission inheritance
      security:
        - BearerAuth: []
      responses:
        "204": { description: Linked }
    delete:
      summary: Remove role inheritance link
      security:
        - BearerAuth: []
      responses:
        "204": { description: Removed }
  /v1/casbin/rules:
    post:
      summary: Create Casbin policy rule
      security:
        - BearerAuth: []
      responses:
        "200": { description: Rule created }
  /v1/webauthn/register/begin:
    post:
      summary: Begin WebAuthn registration (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Creation challenge }
...
  /v1/webauthn/register/finish:
    post:
      summary: Complete WebAuthn registration
      security:
        - BearerAuth: []
      responses:
        "200": { description: Credential registered }
...
  /v1/mfa/verify:
    post:
      summary: Verify TOTP enrollment
      security:
        - BearerAuth: []
      responses:
        "200": { description: MFA enabled }
  /v1/mfa/disable:
    post:
      summary: Disable MFA
      security:
        - BearerAuth: []
      responses:
        "200": { description: MFA disabled }
  /v1/users/{id}/export:
    get:
      summary: DSAR data export (self or admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: User data bundle }
  /v1/users/{id}/privacy:
    delete:
      summary: DSAR erasure (self or admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: User erased }
  /v1/users/{id}/sessions:
    get:
      summary: List user sessions (self or admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Session list }
  /v1/sessions/{id}:
    delete:
      summary: Revoke session (self or admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Session revoked }
  /v1/compliance/status:
    get:
      summary: Compliance posture (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Compliance status }
  /v1/scim-tokens:
    post:
      summary: Create SCIM bearer token (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Token created }
  /v1/webhooks:
    post:
      summary: Register webhook (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Webhook created }
  /v1/idp-configs:
    get:
      summary: List IdP configs (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: IdP config list }
  /v1/idp-configs/{provider}:
    post:
      summary: Upsert IdP config (admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Config saved }
    delete:
      summary: Delete IdP config (admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
...
  /scim/v2/Users:
    get:
      summary: List SCIM users
      security:
        - BearerAuth: []
      responses:
        "200": { description: SCIM list response }
    post:
      summary: Provision SCIM user
      security:
        - BearerAuth: []
      responses:
        "201": { description: User created }
  /scim/v2/Users/{id}:
    get:
      summary: Get SCIM user
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: SCIM user }
    delete:
      summary: Deprovision SCIM user
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
  /scim/v2/Groups:
    get:
      summary: List SCIM groups (roles)
      security:
        - BearerAuth: []
      responses:
        "200": { description: SCIM group list }
  /v1/accounts:
    post:
      summary: Create account (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Account created }
  /v1/accounts/{id}:
    get:
      summary: Get account (admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Account }
  /v1/websites:
    post:
      summary: Create website (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Website created }
  /v1/roles:
    post:
      summary: Create role (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Role created }
  /v1/permissions:
    post:
      summary: Create permission (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Permission created }
  /v1/users/{id}/roles:
    post:
      summary: Assign role to user (admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Assigned }
  /v1/api-keys/{id}:
    delete:
      summary: Revoke API key (admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Revoked }
  /v1/clients/{id}:
    delete:
      summary: Delete OAuth client (admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
  /v1/users:
    get:
      summary: Search users by email prefix (admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: email, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: User search results }
  /v1/audit/events:
    get:
      summary: List audit events (admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
      responses:
        "200": { description: Audit event list }
  /v1/saml/service-providers:
    get:
      summary: List SAML IdP service providers (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Service provider list }
    post:
      summary: Register SAML service provider (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Service provider created }
  /v1/saml/service-providers/{id}:
    delete:
      summary: Delete SAML service provider (admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
...
  /admin/session:
    post:
      summary: Create admin console session (httpOnly cookie)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "204": { description: Session created }
    get:
      summary: Admin session status
      responses:
        "200": { description: Session status }
    delete:
      summary: Destroy admin session
      responses:
        "204": { description: Session cleared }
...